Leave build scripts around in debug mode (#1442015)
We do this for install-time scripts too, so its only logical to do
so with build scripts as well.
Depends on commit 724830018a455f25d8a9be60f7d858968fd6b7a6.

(cherry picked from commit ee33773218d5bda08a32c3775101b8aa211f2b13)

+++ build/build.c
@@ -183,7 +183,7 @@ rpmRC doScript(rpmSpec spec, rpmBuildFlags what, const char *name,
 exit:
     Fclose(fd);
     if (scriptName) {
-	if (rc == RPMRC_OK)
+	if (rc == RPMRC_OK && !rpmIsDebug())
 	    (void) unlink(scriptName);
 	free(scriptName);
     }
